+++ src/vue/layout.ts
@@ -8,6 +8,15 @@ export { default as GkStack } from './components/stack/GkStack.vue'
 export { default as GkContainer } from './components/container/GkContainer.vue'
 export { default as GkDivider } from './components/divider/GkDivider.vue'
 export { default as GkSpinner } from './components/spinner/GkSpinner.vue'
+export { default as GkSkeletonLoader } from './components/feedback/skeleton-loader/GkSkeletonLoader.vue'
+export { gkSkeletonRootTypes } from './components/feedback/skeleton-loader/gk-skeleton-root-types'
+export type { GkSkeletonLoaderType } from './components/feedback/skeleton-loader/gk-skeleton-root-types'
+export {
+  genGkSkeletonStructure,
+  isGkSkeletonLoaderType,
+  wrapGkSkeletonTypes,
+} from './components/feedback/skeleton-loader/gk-skeleton-tree'
+export type { GkSkeletonNode } from './components/feedback/skeleton-loader/gk-skeleton-tree'
 export { default as GkExpansionPanels } from './components/expansion/GkExpansionPanels.vue'
 export { default as GkExpansionPanel } from './components/expansion/GkExpansionPanel.vue'
 export { default as GkExpansionPanelTitle } from './components/expansion/GkExpansionPanelTitle.vue'